% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATD.R
\name{momATD}
\alias{momATD}
\title{Method of moments (MoM) estimation for the quantile-based asymmetric Student's-\eqn{t} distribution.}
\usage{
momATD(y, alpha = NULL)
}
\arguments{
\item{y}{This is a vector of quantiles.}

\item{alpha}{This is the index parameter  \eqn{\alpha}.  If \eqn{\alpha} is unknown, indicate NULL which is the default option. In this case, the sample skewness will be used to estimate \eqn{\alpha}. If \eqn{\alpha} is known, then the value of \eqn{\alpha} has to be specified in the function.}
}
\value{
\code{\link{momATD}} provides the method of moments estimates of the unknown parameters of the distribution.
}
\description{
Parameter estimation in the quantile-based asymmetric Student's-\eqn{t} distribution by using method of moments are discussed in Gijbels et al. (2019a). We here used the first four sample moments to estimate parameter \eqn{\theta=(\mu,\phi,\alpha,\nu)} under the assumption that the first four population moments exist, which needs to assume \eqn{\nu>4}.
}
\examples{
# Example
y=rnorm(100)
momATD(y=y,alpha=0.5) # If alpha is known with alpha=0.5
momATD(y=y) # If alpha is unknown

}
\references{
{
 Gijbels, I., Karim, R. and Verhasselt, A. (2019a). On quantile-based asymmetric family of distributions: properties and inference. \emph{International Statistical Review}, \url{https://doi.org/10.1111/insr.12324}.
}
}
